requirements.txt
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
docutils==0.22.4
filetype==1.2.0
idna==3.11
ImageIO==2.37.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
lazy_loader==0.4
llvmlite==0.46.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
Pygments==2.19.2
PyMatting==1.1.14
pypiwin32==223
pywin32==311
referencing==0.37.0
rembg==2.0.69
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.16.3
tifffile==2025.12.20
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
